requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
mediapipe>=0.10.13
numpy==1.26.0
Pillow==10.0.1